Remove vftr_inittime
All times are now measured wrt the actual beginning of vftr_setup.
Previously, there was a difference between the reference time
of the runtime and the overhead time. This difference is very small
for real applications, but can lead to negative runtimes for very short
runs, where the overhead is larger than the total runtime due to this
difference.